+++ Runtime/TestPlatform.cs
@@ -11,30 +11,29 @@ namespace UnityEngine.TestTools.Graphics
     /// </summary>
     public class TestPlatform
     {
-        private readonly Architecture architecture;
-        private readonly RuntimePlatform platform;
-
         /// <summary>
         /// The OS architecture of the test environment.
         /// </summary>
-        public Architecture Arch => architecture;
+        public Architecture Arch { get; }
+
+        private static readonly Architecture? TestSettingsArchitecture = RuntimeSettings.TryReadSettingsFromFile(RuntimeSettings.FindCommandLineArgument("-testSettingsFile"))?.Architecture;
 
         /// <summary>
         /// The runtime platform value of the test environment.
         /// </summary>
-        public RuntimePlatform Platform => platform;
+        private RuntimePlatform Platform { get; }
 
-        public TestPlatform(RuntimePlatform platform, Architecture architecture)
+        private TestPlatform(RuntimePlatform platform, Architecture architecture)
         {
-            this.platform = platform;
-            this.architecture = architecture;
+            Platform = platform;
+            this.Arch = architecture;
         }
 
         /// <summary>
         /// Returns the current test environment's platform and architecture.
         /// </summary>
         /// <returns></returns>
-        public static TestPlatform GetCurrent()
+        public static TestPlatform GetTarget()
         {
             var currentPlatform = Application.platform;
             var currentArchitecture = System.Runtime.InteropServices.RuntimeInformation.OSArchitecture;
@@ -45,13 +44,15 @@ namespace UnityEngine.TestTools.Graphics
                 currentArchitecture = System.Runtime.InteropServices.Architecture.Arm64;
             #endif
 
-            return new TestPlatform(currentPlatform, currentArchitecture);
+            var architecture = TestSettingsArchitecture ?? currentArchitecture;
+
+            return new TestPlatform(currentPlatform, architecture);
         }
 
         /// <summary>
         /// Converts the TestPlatform to a unique string value.
         /// </summary>
         /// <returns></returns>
-        public override string ToString() => platform.ToUniqueString(architecture);
+        public override string ToString() => Platform.ToUniqueString(Arch);
     }
 }